airflow/requirements_docker.txt
requests
openai 
scrapy 
beautifulsoup4 
loguru 
newspaper3k 
lxml-html-clean
numpy
pandas
wordcloud